Sign method now works using concept. Update sign method everywhere. Sign tests partially work.
entry is now considered to be valid if both Key and Value are valid. Previously, an invalid
value was considered to be that which is obtained if Value is constructed with the default
constructor. However, this proved to be inconvenient. It is now possible to use a map with a
value that is valid if default constructed. If you do this, the [] operator will return the
default constructed value as usual if a key is not in the map, but now you don't know if it
was in the map or not!

The contains method has been changed to return a pointer to the value if it exists and nullptr
if it doen't. This is very convenient because the pointer can be cast to bool and used the
same way as before but you can also find the element you asked about if you want.

Permutations multiplication has been partially fixed.